+++ smartsheet/models/__init__.py
@@ -43,6 +43,7 @@ from .copy_or_move_row_result import CopyOrMoveRowResult
 from .criteria import Criteria
 from .cross_sheet_reference import CrossSheetReference
 from .currency import Currency
+from .data_classification import DataClassification
 from .date_object_value import DateObjectValue
 from .discussion import Discussion
 from .downloaded_file import DownloadedFile