+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [x.x.x] - Unreleased
 
+### Added
+
+- Support for PUT /sheets/{sheetId}/dataclassification endpoint (Set Data Classification)
+- Support for DELETE /sheets/{sheetId}/dataclassification endpoint (Remove Data Classification)
+- Added `data_classification` field to Sheet model
+
 ## [4.4.0] - 2026-08-12
 
 ### Added